requirements.txt
requests
beautifulsoup4
pymysql
python-dotenv
yfinance
selenium
webdriver-manager
flask
schedule